Splited the clint and server component for login and register page
