+++ src/app/register/page.tsx
@@ -1,77 +1,11 @@
-"use client";
+// Registration Page (Server Component Wrapper)
+// -------------------------------------------
+// Provides static shell & mounts interactive RegisterForm client component.
+// Future server logic: redirect if authenticated, fetch invite codes, etc.
 
-// Registration Page (Collaborative Notes)
-// ---------------------------------------
-// PURPOSE:
-//   Basic new user sign-up form (name, email, password, confirm) for RentEase.
-//   Placeholder until real backend / auth provider is integrated.
-//
-// TEAM EXTENSION POINTS:
-//   - AUTH_SUBMIT: Replace mock submit with real API / NextAuth signUp (if custom flow).
-//   - FORM_VALIDATION: Integrate schema validation (Zod/Yup) & show richer messages.
-//   - PASSWORD_RULES: Enforce stronger rules (length, complexity) & realtime feedback.
-//   - SOCIAL_PROVIDERS: Add OAuth sign-up buttons.
-//   - ERROR_UI: Replace inline error text with toast system.
-//
-// DESIGN NOTES:
-//   - Mirrors look & feel of Login page for consistency.
-//   - Minimal CSS via Tailwind / DaisyUI utilities only.
-//   - Accessible form controls, labels, and error region.
-//
-// FUTURE IDEAS:
-//   - Add terms / privacy consent checkbox.
-//   - Email verification flow.
-//
-import { FormEvent, useState } from "react";
-import Link from "next/link";
-
-interface FormState {
-  name: string;
-  email: string;
-  password: string;
-  confirm: string;
-}
+import { RegisterForm } from "@/components/auth/RegisterForm";
 
 export default function RegisterPage() {
-  const [form, setForm] = useState<FormState>({ name: "", email: "", password: "", confirm: "" });
-  const [loading, setLoading] = useState(false);
-  const [error, setError] = useState<string | null>(null);
-  const [success, setSuccess] = useState(false);
-
-  const handleSubmit = async (e: FormEvent) => {
-    e.preventDefault();
-    setError(null);
-    setSuccess(false);
-
-    // (FORM_VALIDATION) Simple synchronous checks.
-    if (!form.name || !form.email || !form.password) {
-      setError("All fields are required");
-      return;
-    }
-    if (form.password.length < 6) {
-      setError("Password must be at least 6 characters");
-      return;
-    }
-    if (form.password !== form.confirm) {
-      setError("Passwords do not match");
-      return;
-    }
-
-    try {
-      setLoading(true);
-      // (AUTH_SUBMIT) Replace with real POST /api/auth/register.
-      await new Promise((res) => setTimeout(res, 900));
-  console.log("REGISTER_SUBMIT", form);
-      setSuccess(true);
-      // Optionally redirect after slight delay.
-      // router.push('/login');
-    } catch {
-      setError("Registration failed (mock)");
-    } finally {
-      setLoading(false);
-    }
-  };
-
   return (
     <main className="min-h-[calc(100vh-4rem)] flex items-center justify-center px-4 py-10">
       <div className="w-full max-w-md">
@@ -79,97 +13,7 @@ export default function RegisterPage() {
           <div className="card-body">
             <h1 className="text-2xl font-semibold tracking-tight mb-1">Create an account</h1>
             <p className="text-base-content/70 text-sm mb-4">Join RentEase and start listing or finding rentals.</p>
-
-            <form onSubmit={handleSubmit} className="space-y-4" noValidate>
-              <div className="form-control">
-                <label className="label" htmlFor="name">
-                  <span className="label-text font-medium">Full name</span>
-                </label>
-                <input
-                  id="name"
-                  type="text"
-                  className="input input-bordered w-full"
-                  placeholder="Jane Doe"
-                  value={form.name}
-                  onChange={(e) => setForm({ ...form, name: e.target.value })}
-                  required
-                />
-              </div>
-
-              <div className="form-control">
-                <label className="label" htmlFor="email">
-                  <span className="label-text font-medium">Email</span>
-                </label>
-                <input
-                  id="email"
-                  type="email"
-                  autoComplete="email"
-                  className="input input-bordered w-full"
-                  placeholder="you@example.com"
-                  value={form.email}
-                  onChange={(e) => setForm({ ...form, email: e.target.value })}
-                  required
-                />
-              </div>
-
-              <div className="form-control">
-                <label className="label" htmlFor="password">
-                  <span className="label-text font-medium">Password</span>
-                </label>
-                <input
-                  id="password"
-                  type="password"
-                  autoComplete="new-password"
-                  className="input input-bordered w-full"
-                  placeholder="••••••••"
-                  value={form.password}
-                  onChange={(e) => setForm({ ...form, password: e.target.value })}
-                  required
-                  minLength={6}
-                />
-              </div>
-
-              <div className="form-control">
-                <label className="label" htmlFor="confirm">
-                  <span className="label-text font-medium">Confirm password</span>
-                </label>
-                <input
-                  id="confirm"
-                  type="password"
-                  autoComplete="new-password"
-                  className="input input-bordered w-full"
-                  placeholder="••••••••"
-                  value={form.confirm}
-                  onChange={(e) => setForm({ ...form, confirm: e.target.value })}
-                  required
-                  minLength={6}
-                />
-              </div>
-
-              {/* Feedback region (ERROR_UI) */}
-              <div aria-live="polite" className="min-h-5 text-sm">
-                {error && <span className="text-error">{error}</span>}
-                {success && !error && (
-                  <span className="text-success">Account created (mock). You can now login.</span>
-                )}
-              </div>
-
-              <button
-                type="submit"
-                className="btn btn-primary w-full"
-                disabled={loading}
-              >
-                {loading ? <span className="loading loading-spinner loading-sm" /> : "Create account"}
-              </button>
-            </form>
-
-            {/* (SOCIAL_PROVIDERS) Insert provider buttons here */}
-            <div className="mt-6 text-center text-sm">
-              <span className="text-base-content/70">Already have an account? </span>
-              <Link href="/login" className="link link-primary font-medium">
-                Login
-              </Link>
-            </div>
+            <RegisterForm />
           </div>
         </div>
       </div>
